Avoid negative numbers in json.

+++ src/modules/module.cpp
@@ -627,7 +627,7 @@ QJsonObject MtbModule::dvRepr(uint8_t dvi, const std::vector<uint8_t> &data) con
 		case Mtb::DVCommon::MtbBusBadCrc:
 		case Mtb::DVCommon::MtbBusSent:
 			if (data.size() == 4)
-				return {{Mtb::DVCommonToStr(dvi), static_cast<int>(pack<uint32_t>(data))}};
+				return {{Mtb::DVCommonToStr(dvi), static_cast<qint64>(pack<uint32_t>(data))}};
 			break;
 	}
